refactor: move unit tests into package as tested function
