+++ packages/backend/api/getId/getId.go
@@ -1,4 +1,4 @@
-package api
+package getId
 
 import (
 	"fmt"
