+++ packages/backend/api/getRecentTicketInspectorInfo/test_getRecentTicketInspectorInfo/IdToCoordinates_test.go
@@ -1,4 +1,4 @@
-package api_test
+package getRecentTicketInspectorInfo_test
 
 import (
 	"fmt"
