+++ packages/backend/api/getStationDistance/test_getStationDistance/GetNearestStation_test.go
@@ -1,4 +1,4 @@
-package api_test
+package test_getStationDistance
 
 import (
 	"testing"
