+++ packages/backend/api/getStationDistance/test_getStationDistance/GetStationDistance_test.go
@@ -1,4 +1,4 @@
-package api_test
+package test_getStationDistance
 
 import (
 	"net/http"
