+++ packages/backend/main.go
@@ -4,9 +4,13 @@ import (
 	"net/http"
 
 	"github.com/FreiFahren/backend/Rstats"
-	"github.com/FreiFahren/backend/api"
+	"github.com/FreiFahren/backend/api/getAllStationsAndLines"
+	"github.com/FreiFahren/backend/api/getId"
 	"github.com/FreiFahren/backend/api/getRecentTicketInspectorInfo"
+	"github.com/FreiFahren/backend/api/getSegmentColors"
 	"github.com/FreiFahren/backend/api/getStationDistance"
+	"github.com/FreiFahren/backend/api/getStationName"
+	"github.com/FreiFahren/backend/api/getStats"
 	"github.com/FreiFahren/backend/api/postInspector"
 	"github.com/FreiFahren/backend/data"
 	"github.com/FreiFahren/backend/database"
@@ -115,22 +119,22 @@ func main() {
 	apiHOST.GET("/basics/recent", getRecentTicketInspectorInfo.GetRecentTicketInspectorInfo)
 
 	// Return the name for given id
-	apiHOST.GET("/data/station", api.GetStationName)
+	apiHOST.GET("/data/station", getStationName.GetStationName)
 
 	// Return all stations with their id
-	apiHOST.GET("/data/list", api.GetAllStationsAndLines)
+	apiHOST.GET("/data/list", getAllStationsAndLines.GetAllStationsAndLines)
 
 	// Return the id for given name
-	apiHOST.GET("/data/id", api.GetStationId)
+	apiHOST.GET("/data/id", getId.GetStationId)
 
 	// Return the distance between two stations
 	apiHOST.GET("/transit/distance", getStationDistance.GetStationDistance)
 
 	// Get usage statistics
-	apiHOST.GET("/statistics/stats", api.GetStats)
+	apiHOST.GET("/statistics/stats", getStats.GetStats)
 
 	// Get the list of highlighted segments and their colors
-	apiHOST.GET("/risk-prediction/getSegmentColors", api.GetSegmentColors)
+	apiHOST.GET("/risk-prediction/getSegmentColors", getSegmentColors.GetSegmentColors)
 
 	apiHOST.Start(":8080")
